chore: rename to Microsoft.Playwright.Xunit
