+++ src/Playwright.Examples/Program.cs
@@ -29,4 +29,3 @@ await using var browser = await playwright.Chromium.LaunchAsync();
 var page = await browser.NewPageAsync();
 await page.GotoAsync("https://playwright.dev/dotnet");
 await page.ScreenshotAsync(new PageScreenshotOptions { Path = "screenshot.png" });
-
