+++ src/Playwright.Xunit/PlaywrightTest.cs
@@ -25,7 +25,7 @@
 using System.Threading.Tasks;
 using Microsoft.Playwright.TestAdapter;
 
-namespace Microsoft.Playwright.xunit;
+namespace Microsoft.Playwright.Xunit;
 
 public class PlaywrightTest : WorkerAwareTest
 {
